First commit to Community Plus
Community Plus is a web application where users can post their problems or their need of any service or assistance.  The users will have to select a domain (ex. Medical, Automobile, Household, etc.) in which they are willing to provide help while registering themselves to Community Plus. When a user publishes a new post, the backend logic of Community Plus sends an email about this new post as a notification to all other users who has the same help domain preference with that of the domain of the subject of the post. Community Plus uses a chronological feed to display all the posts with the most recent posts being at the top.